refactor specs
move TestJob et al into supporting role
add new supporting classes
