+++ spec/support/test_job_task.rb
@@ -1,0 +1,2 @@
+class TestJobTask < Taskinator::Task::Job
+end
