+++ spec/support/test_step_task.rb
@@ -1,0 +1,2 @@
+class TestStepTask < Taskinator::Task::Step
+end
