+++ spec/support/test_subprocess_task.rb
@@ -1,0 +1,2 @@
+class TestSubProcessTask < Taskinator::Task::SubProcess
+end
