+++ spec/taskinator/task_spec.rb
@@ -331,37 +331,6 @@ describe Taskinator::Task do
 
   describe Taskinator::Task::Job do
 
-    module TestJob
-      def self.perform(*args)
-      end
-    end
-
-    class TestJobClass
-      def perform(*args)
-      end
-    end
-
-    module TestJobModule
-      def self.perform(*args)
-      end
-    end
-
-    class TestJobClassNoArgs
-      def perform
-      end
-    end
-
-    module TestJobModuleNoArgs
-      def self.perform
-      end
-    end
-
-    module TestJobError
-      def self.perform
-        raise ArgumentError
-      end
-    end
-
     subject { Taskinator::Task.define_job_task(process, TestJob, [1, {:a => 1, :b => 2}]) }
 
     it_should_behave_like "a task", Taskinator::Task::Job
